Refactor by removing the scroll model out of the root file
I consider this root file to be the view layer, which shouldn't have too 
much knowledge of the model. The model can be re-used for multiple views 
without needing to change too much. In short, some basic refactoring to 
improve the architecture. No big deal.
