+++ index.js
@@ -1,11 +1,9 @@
 /* global gtag */
 
-import { InfiniteScroller, IntegerIterator, title as getTitle, physicalLocationFromRef, urlToRef } from './src'
+import { InfiniteScroller, urlToRef, scrollsByKey } from './src'
 import Page from './components/Page'
 import ParshaPicker from './components/ParshaPicker'
 import utils from './components/utils'
-import pageTitles from './build/page-titles.json'
-import holydays from './build/holydays.json'
 
 const { htmlToElement, whenKey, purgeNode } = utils
 
@@ -41,71 +39,9 @@ const scrollToLine = ({ node, lineIndex }) => {
   book.scrollTop = line.offsetTop + (line.offsetHeight / 2) - (book.offsetHeight / 2)
 }
 
-const scrollsByKey = () => ({
-  'torah': TorahScroll,
-  'esther': EstherScroll,
-  ...Object.keys(holydays).reduce((result, holydayKey) => {
-    const HolydayScroll = {
-      new: ({ startingAtRef }) => {
-        return Scroll.new({
-          scroll: holydayKey,
-          makePath: n => `/build/pages/${holydayKey}/${n}.json`,
-          makeTitle: n => holydays[holydayKey].he,
-          startingAtRef
-        })
-      }
-    }
-    return { ...result, [holydayKey]: HolydayScroll }
-  }, {})
-})
-
-const Scroll = {
-  new: ({ scroll, makePath, makeTitle, startingAtRef = { b: 1, c: 1, v: 1 } }) => {
-    const { pageNumber, lineNumber } = physicalLocationFromRef({ ref: startingAtRef, scroll })
-
-    const iterator = IntegerIterator.new({ startingAt: pageNumber })
-
-    return {
-      scrollName: scroll,
-      fetchPrevious: () => {
-        const n = iterator.previous()
-        if (n <= 0) return Promise.resolve()
-        return fetchPage({ path: makePath(n), title: makeTitle(n) })
-      },
-      fetchNext: () => {
-        const n = iterator.next()
-        return fetchPage({ path: makePath(n), title: makeTitle(n) })
-      },
-      startingLineNumber: lineNumber
-    }
-  }
-}
-
-const TorahScroll = {
-  new: ({ startingAtRef }) => {
-    return Scroll.new({
-      scroll: 'torah',
-      makePath: n => `/build/pages/torah/${n}.json`,
-      makeTitle: n => getTitle(pageTitles[n - 1]),
-      startingAtRef
-    })
-  }
-}
-
-const EstherScroll = {
-  new: ({ startingAtRef }) => {
-    return Scroll.new({
-      scroll: 'esther',
-      makePath: n => `/build/pages/esther/${n}.json`,
-      makeTitle: n => 'אסתר',
-      startingAtRef
-    })
-  }
-}
-
 const app = {
   jumpTo: ({ ref }) => {
-    scroll = scrollsByKey()[ref.scroll].new({ startingAtRef: ref })
+    scroll = scrollsByKey[ref.scroll].new({ startingAtRef: ref })
 
     purgeNode(document.querySelector('[data-target-id="tikkun-book"]'))
 
@@ -262,13 +198,6 @@ const renderPage = ({ insertStrategy: insert }) => ({ content, title }) => {
 
 const renderPrevious = renderPage({ insertStrategy: insertBefore })
 const renderNext = renderPage({ insertStrategy: insertAfter })
-
-const fetchPage = ({ path, title }) => window.fetch(path)
-  .then((res) => res.json())
-  .then((page) => ({ content: page, title }))
-  .catch((err) => {
-    console.error(err)
-  })
 
 const debounce = (callback, delay) => {
   let timeout
