+++ src/index.js
@@ -1,5 +1,6 @@
 import InfiniteScroller from './infinite-scroller'
 import IntegerIterator from './integer-iterator'
+import scrollsByKey from './scrolls-by-key'
 const hebrewNumeralFromInteger = require('./hebrew-numeral')
 const textFilter = require('./text-filter')
 const displayRange = require('./display-range')
@@ -7,4 +8,4 @@ const title = require('./title')
 const { physicalLocationFromRef } = require('./location')
 const urlToRef = require('./url-to-ref')
 
-export { textFilter, hebrewNumeralFromInteger, displayRange, InfiniteScroller, IntegerIterator, title, physicalLocationFromRef, urlToRef }
+export { textFilter, hebrewNumeralFromInteger, displayRange, InfiniteScroller, IntegerIterator, title, physicalLocationFromRef, urlToRef, scrollsByKey }
